requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex
_openmp_mutex
backcall
blas
bzip2
ca-certificates
certifi
cffi
cudatoolkit
decorator
ffmpeg
freetype
future
# fvcore
gmp
gnutls
intel-openmp
# iopath
ipython
ipython_genutils
jedi
joblib
jpeg
lame
lcms2
ld_impl_linux-64
# libblas
# libcblas
libedit
libffi
libgcc-ng
libgfortran-ng
libgfortran5
libgomp
libiconv
# liblapack
libpng
libstdcxx-ng
libtiff
libuv
llvm-openmp
lz4-c
mkl
mkl-service
mkl_fft
mkl_random
ncurses
nettle
ninja
numpy
numpy-base
# nvidiacub
olefile
openh264
openssl
parso
pexpect
pickleshare
pillow
pip
portalocker
prompt-toolkit
ptyprocess
pycparser
pygments
pyyaml
readline
scikit-learn
scipy
setuptools
six
sqlite
tabulate
termcolor
threadpoolctl
tk
torchvision
tqdm
traitlets
# trimesh
typing_extensions
wcwidth
wheel
xz
yacs
yaml
zlib
zstd
